apps/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.3.0
motor==3.4.0          # Async MongoDB driver
pymongo==4.7.2
httpx==0.27.0
minio==7.2.7
redis==5.0.4
aioredis==2.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
bson==0.5.10